input: lsm6ds3: improving polling performance and timestamps
Polling with a fixed timer delay leads to a lower than expected ODR
(output data rate) because the effective sampling period is
requested_period + time_spent_in_poll. Use hrtimer_forward to advance
the timer to the next period past now.

Remove an old bug-fix that attempted to improve sampling performance in
the case where processing took longer than a one period by expiring the
timer immediately. This bug-fix is problematic because we don't know if
the sensor has had enough time to produce a new sample since the one we
just read.

Take the sensor event timestamp before reading from the sensor because
it is likely to be closer to the time at which the sample was recorded
on the sensor.  It's not possible to accurately synchronize timestamps
in polled mode.

SAMBAR-156

Change-Id: I26af361ae065e9f07584628673747f728c95dbe1
Signed-off-by: Khalid Zubair <kzubair@cyngn.com>