defconfig: sambar: minimize cyanogenmod_sambar_defconfig
Generate a minimized defconfig by running "make menuconfig" and
saving, without actually making any changes. This strips out
redundant lines that specify configurations which are default or
implicit based on other options, and reduces the potential for
conflicts when future changes to the defconfig are made.

As an added bonus, this avoids diffs appearing in the defconfig
as a result of simply building the kernel.

No functional change is expected due to this change.

SAMBAR-1000
Change-Id: Ieb85e5f2cb850ae24d34b86a5b7e695e5b7500b8
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>