defconfig: introduce cyanogenmod_kerneldebug_config
cyanogenmod_kerneldebug_defconfig is a defconfig fragment containing
debug related features which may not be appropriate to enable in
production builds due to their impact on performance or power.

This is not a full defconfig, just a list of deltas. It's meant
to be applied to a base defconfig usesing merge-config.sh.

SAMBAR-1000
Change-Id: I803ae6dcd83b1122d5c360a12825e0e36204074c
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>