ASoC: msm: qdsp6v2: Don't send TruMedia commands to DSP if COPP is resetting
This avoids memory scribbling when update_params_ptr is written to
in srs_trumedia_open() despite the memory it points to having already
been freed.

Although this scenario should not be allowed to arise in the first
place, this added check avoids potential instability if it does.

This avoids the scribbling crash reported in SAMBAR-988, for which a
fix for the COPP reset is still under discussion.

Change-Id: I70843144673ddf35650a6f5796f1f3baea633ca3
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>