Revert "ASoC: soc: msm: dynamic change port id for speaker and headset"
This reverts commit 4e6e4cd0d13b764e6eaa14dcedd63831f4588871.

The userspace libsrsprocessing.so library now supports the "SRS
TruMedia QUAT_MI2S" mixer control, and selects it during loudspeaker
playback. We no longer need to work around this by hacking the port
IDs in the kernel.

The removed code was broken anyway, since the check of the MI2S
clock's enable state could race with that clock being enabled,
resulting in the  SLIMBUS_0_RX port being selected unintentionally.
This was observed to lead to DSP command failures, which further
lead to kernel panics in some instances.

SAMBAR-988
Change-Id: I3e81e48f6be08e4dc52a2ad537cc1426afa75c96
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>