pm9x: updated dsi-on-command sequences for sharp panels
Update the sequences to match more closely what's recommended in the
panel documentation and net some latency improvements as a result.

The panel documentation recommends that any custom panel configuration
be done between the command 0x2A and the command 0xB0. Currently, the
mdss-livedisplay-cabc-cmd sequence doesn't execute until much later.
Move it back to be in line with the recommendations.

The latency improvement comes as a welcomed side-effect of this. With
this change, no command occurs after command 0x11 (which has a large
120 ms delay associated with it). The DSI host driver is smart enough
to notice this and will not insert the delay unless another DSI command
comes in before 120 ms have passed. Since this doesn't seem to happen
in practice, a full 120 ms latency improvement is realized.

While we're at at, also drop command 0x35, which is used to enable
tearing effect (TE) control. This is a standardized MIPI DSI command
which mdss_dsi.c will send as appropriately anyway. It does not need
to be duplicated in the dsi-on-command sequence.

Change-Id: Ia5e8b51717ed4f575a5c64b7536dc3134b6e1cf8
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>
(cherry picked from commit a0bfbfe3853e323c74ccb77f22da12d5e92724b1)