input: hsppad038: fix potential race with work cancellation
There is a small race potential when the hsppad_polling() work is
cancelled that may result in rare unexpected wake-ups. Here's the
problematic order of events:

 1) hsppad_suspend() acquires the lock and calls hsppad_measure_stop()
    with it held.
 2) hsppad_polling() work starts running
 3) hsppad_measure_stop() calls cancel_delayed_work()
 4) hsppad_polling() waits for hsppad_measure_stop() to return and
    the lock to be released, so that it may acquire the lock
 5) hsppad_polling() acquires the lock and proceeds to call
    schedule_delayed_work().

If this happens, the work will not be cancelled as desired and will
continue to reschedule itself. Fix this by cancelling the work
with cancel_delayed_work_sync(). Clean up the locking so that this
can be done without deadlocking.

SAMBAR-1217
Change-Id: I27069179d8c36da17d74c3ea03449fa3a3622cae
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>