ASoC: msm8994: remove wait for external amp power-on
The presence of the MI2S clock is not all that's required to power-on
Sambar's external amp. The amp's PWDN control bit must also be cleared,
and that's done in the amplifier's userspace driver.

This means the delay added after MI2S clock enablement is not in the
right spot. Remove it. The correct delay has been added in userspace.

Issue: SAMBAR-1210
Change-Id: I1cff40cec179d9eff5cb8a0995ac1f7b20b895a7
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>